Make the settings module configurable by ENV variables
Uses common conventions like `DATABASE_URL` and `REDIS_URL`. Also can
set `REDIS_PROVIDER` to the name of a URL variable to be configured by,
similar to Sidekiq. This is useful for oddly-named Heroic providers.
Other vars are named the same as the equivalent settings constants.
Move the thumbnail metadata cache to the database for easier
portability.
Remove the local settings handling